+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Bug #15: Update copyright year to `2024` in `LICENSE` file (@terabytesoftw)
 - Bug #16: Update image alt text from `Yii Framework` to `UI Awesome` (@terabytesoftw)
+- Bug #17: Enhance support for `Stringable` interface types across `BaseCssClass`, `BaseEncode`, `BaseEnum`, `BaseValidator` classes and update related tests (@terabytesoftw)
 
 ## 0.4.0 December 17, 2025